:truck: Move upload helper into separate helpers folder
